Reordered the Bridgehead restart instructions
Ordered them so that first a stop is performed, then a start. This
reflects what administrators will be doing in real life. Also included a
git pull for the Bridgehead, since this will normally also be part of
the regular administrative tasks.

Also removed a bunch of instructions for the installation, since these
steps are now performed automatically by the bridgehead script.