Got systemctl restart to work locally
There were problems on my Ubuntu VM with the target run level that I was
using, so I went for the default. Now, if I restart my VM, the Bridgehead
is successfully started automatically.

Also added some status-generating commands to the run.sh and stop.sh
scripts, so that it looks like they are doing something.