Control import from Directory, improve README
Two new environment variables have been introduced, to allow a user to control
which data (if any) is imported from the Directory during a sync operation.
This was considered necessary, because recent changes to Directory sync have
greatly expanded the data that can be imported.

This was documented in the README, and the oportunity was taken to improve the
existing documentation, especially by peoviding a list of all available
environment variables for Directory sync.