Directory sync: token login and cron change
The Directory team have requested that we allow token login to the
Directory, where a user uses LSAAI credentials to obtain a token from
the Directory, and then uses this to authenticate Directory sync. This
has been implemented via an environment variable, in an analogous way to
the already existing username/password method.

The default start time for the Directory sync has been shifted to 22:30,
to prevent conflicts with the Bridgehead auto-update.

Relevant changes have been made to the documentation.